+++ core/src/main/java/org/dynmapblockscan/core/model/ElementRotation.java
@@ -5,9 +5,9 @@ import java.util.Arrays;
 import org.dynmapblockscan.core.model.ElementRotation;
 
 public class ElementRotation {
-	public float[] origin = { 8.0f, 8.0f, 8.0f };
+	public double[] origin = { 8.0f, 8.0f, 8.0f };
 	public String axis = "x";
-	public float angle = 0.0f;
+	public double angle = 0.0f;
 	public boolean rescale = false;
 	
 	public ElementRotation() {}
